requirements.txt
Jinja2==3.1.0
watchdog==5.0.0
python-frontmatter==1.1.0
mistune==3.0.2